From Steven Legg:
The Attestation type contains a bunch of references to types defined elsewhere. If I included
those other modules as-is then that would draw in other imports and before you know it you
have most of the X.500 modules to content with. To avoid that I created cut down and
simplified versions of those other modules with just the necessary definitions. The complete
collection of ASN.1 modules is in the attached attestation.asn file.

I also made Attestation a parameterized type so that users of the definition would not have
to modify the Attestation module to provide a DataObject. An example of usage is in the
MyAttestation module.